+++ src/components/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import BooksList from '../containers/BooksList';
+import BooksForm from '../containers/BooksForm';
+
+const App = () => (
+  <div className="App">
+    <BooksList />
+    <BooksForm />
+  </div>
+);
+
+export default App;
