+++ src/containers/BooksList.js
@@ -2,7 +2,9 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Book from '../components/Book';
-import { removeBook, getRandomBooks } from '../redux/actions';
+import { removeBook, getRandomBooks, changeFilter } from '../redux/actions';
+import CategoryFilter from '../components/CategoryFilter';
+import Header from '../components/Header';
 import bookCategories from '../config';
 
 class BooksList extends Component {
@@ -16,17 +18,27 @@ class BooksList extends Component {
     removeBook(book);
   };
 
+  handleFilterChange = (filter) => {
+    const { changeFilter } = this.props;
+    changeFilter(filter);
+  };
+
   render() {
     const { books, filter } = this.props;
     return (
       <>
+        <Header>
+          <CategoryFilter
+            filter={filter}
+            handleFilterChange={this.handleFilterChange}
+          />
+        </Header>
         <div className="BookList">
           {books.length === 0 ? (
             <div>Loading...</div>
           ) : (
             books
               .filter((book) => filter === book.category || filter === 'All')
-              .reverse()
               .map((book) => (
                 <Book
                   key={book.id}
@@ -52,6 +64,7 @@ BooksList.propTypes = {
   filter: PropTypes.string.isRequired,
   removeBook: PropTypes.func.isRequired,
   getRandomBooks: PropTypes.func.isRequired,
+  changeFilter: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = ({ books, filter }) => ({
@@ -61,6 +74,7 @@ const mapStateToProps = ({ books, filter }) => ({
 
 const mapDispatchToProps = {
   removeBook,
+  changeFilter,
   getRandomBooks,
 };
