+++ src/redux/actions/index.js
@@ -1,5 +1,3 @@
-import { fetchBooksByCategory } from '../../api/googleBooksAPI';
-
 const actionTypes = {
   CREATE_BOOK: 'CREATE_BOOK',
   REMOVE_BOOK: 'REMOVE_BOOK',
@@ -23,6 +21,25 @@ const changeFilter = (filter) => ({
 });
 
 const getRandomBooks = (categories) => {
+  const fetchBooks = async (searchparams) => {
+    const response = await fetch(
+      `https://www.googleapis.com/books/v1/volumes?${searchparams}`
+    );
+    if (!response.ok) {
+      throw new Error(`Something went wrong ${response.statusText}`);
+    }
+    return response.json();
+  };
+
+  const fetchBooksByCategory = (category) => {
+    const searchparams = new URLSearchParams({
+      q: `subject:${category}`,
+      maxResults: 40,
+      fields: 'items(id,volumeInfo(title,imageLinks/smallThumbnail))',
+    });
+    return fetchBooks(searchparams);
+  };
+
   const MAX_BOOKS_PER_CATEGORY = 3;
 
   function sample(array, size) {
